requirements.txt
requests
django
djangorestframework
python-decouple
